api/requirements.txt
django #framework
gunicorn #deployment
drf-yasg #swagger
whitenoise #staticfiles
django-cors-headers #cors
djangorestframework #rest
djangorestframework-simplejwt #JWT